move functionality to correct locations
